RAG-Task/requirements.txt
langchain
pypdf2
openai
faiss-cpu
python-dotenv
streamlit
langchain_text_splitters
langchain-openai
langchain-community
langchain-core
langchain-classic
pdf2image
pytesseract
chromadb
langfuse
litellm
langchain-openai
langchain-community